integrate Error Enums into course plans
