+++ flaskr/api/course_plans.py
@@ -1,8 +1,8 @@
-from bson import ObjectId
 from flask import Blueprint
 from flask_pydantic import validate  # type: ignore
 
 from flaskr.api.auth_guard import auth_guard
+from flaskr.api.errors import ResponseError
 from flaskr.api.reqmodels import (
     CoursePlanCreateRequestModel,
     CoursePlanUpdateRequestModel,
@@ -21,6 +21,7 @@ from flaskr.db.course_plans import (
 )
 from flaskr.db.models import CoursePlanRead, CoursePlanUpdate, SemesterPlanRead, User
 from flaskr.db.semester_plans import get_semester_plans_by_course_plan
+from flaskr.utils import PydanticObjectId
 
 route = Blueprint("course-plans", __name__, url_prefix="/course-plans")
 
@@ -40,17 +41,12 @@ def read_all(user: User):
 @route.route("/<course_plan_id>", methods=["GET"])
 @auth_guard
 @validate(response_by_alias=True)
-def read_one(course_plan_id: str, user: User):
+def read_one(course_plan_id: PydanticObjectId, user: User):
     assert user.id is not None, "User ID will never be None here"
-    if not ObjectId.is_valid(course_plan_id):
-        return (
-            CoursePlanResponseModel(status="ERROR", error="Malformed ID"),
-            400,
-        )
-    course_plan = get_course_plan(ObjectId(course_plan_id), user.id)
+    course_plan = get_course_plan(course_plan_id, user.id)
     if not course_plan:
         return (
-            CoursePlanResponseModel(status="ERROR", error="Not found"),
+            CoursePlanResponseModel(status="ERROR", error=ResponseError.NotFound),
             404,
         )
     course_plan_read = CoursePlanRead.model_validate(course_plan.model_dump())
@@ -81,7 +77,10 @@ def create(body: CoursePlanCreateRequestModel, user: User):
         description=body.description, name=body.name, user_id=user.id
     )
     if not res:
-        return CoursePlanResponseModel(status="ERROR", error="Unknown error"), 500
+        return (
+            CoursePlanResponseModel(status="ERROR", error=ResponseError.InternalError),
+            500,
+        )
     course_plan_read = CoursePlanRead.model_validate(res.model_dump())
     return (
         CoursePlanResponseModel(status="OK", data=course_plan_read),
@@ -92,22 +91,22 @@ def create(body: CoursePlanCreateRequestModel, user: User):
 @route.route("/<course_plan_id>", methods=["PATCH"])
 @auth_guard
 @validate(response_by_alias=True)
-def update(course_plan_id: str, body: CoursePlanUpdateRequestModel, user: User):
+def update(
+    course_plan_id: PydanticObjectId, body: CoursePlanUpdateRequestModel, user: User
+):
     assert user.id is not None, "User ID will never be None here"
-    if not ObjectId.is_valid(course_plan_id):
-        return (
-            CoursePlanResponseModel(status="ERROR", error="Malformed course plan ID"),
-            400,
-        )
     res = update_course_plan(
-        course_plan_id=ObjectId(course_plan_id),
+        course_plan_id=course_plan_id,
         course_plan_update=CoursePlanUpdate.model_validate(
             body.model_dump(exclude_none=True)
         ),
         user_id=user.id,
     )
     if not res:
-        return CoursePlanResponseModel(status="ERROR", error="Unknown error"), 500
+        return (
+            CoursePlanResponseModel(status="ERROR", error=ResponseError.InternalError),
+            500,
+        )
     course_plan_read = CoursePlanRead.model_validate(res.model_dump())
     return CoursePlanResponseModel(status="OK", data=course_plan_read), 200
 
@@ -115,14 +114,12 @@ def update(course_plan_id: str, body: CoursePlanUpdateRequestModel, user: User):
 @route.route("/<course_plan_id>", methods=["DELETE"])
 @auth_guard
 @validate(response_by_alias=True)
-def delete(course_plan_id: str, user: User):
+def delete(course_plan_id: PydanticObjectId, user: User):
     assert user.id is not None, "User ID will never be None here"
-    if not ObjectId.is_valid(course_plan_id):
-        return (
-            CoursePlanResponseModel(status="ERROR", error="Malformed course plan ID"),
-            400,
-        )
-    res = delete_course_plan(course_plan_id=ObjectId(course_plan_id), user_id=user.id)
+    res = delete_course_plan(course_plan_id=course_plan_id, user_id=user.id)
     if not res:
-        return CoursePlanResponseModel(status="ERROR", error="Not found"), 404
+        return (
+            CoursePlanResponseModel(status="ERROR", error=ResponseError.NotFound),
+            404,
+        )
     return CoursePlanResponseModel(status="OK"), 200
