+++ flaskr/db/models.py
@@ -1,43 +1,9 @@
 from datetime import datetime, timezone
-from typing import Any, ClassVar, Optional
+from typing import ClassVar, Optional
 
-from bson import ObjectId
 from pydantic import BaseModel, ConfigDict, Field
-from pydantic_core import core_schema
-from typing_extensions import Annotated
 
-
-# copy pasted from pydantic_mongo.fields
-class ObjectIdAnnotation:
-    @classmethod
-    def __get_pydantic_core_schema__(
-        cls, _source_type: Any, _handler: Any
-    ) -> core_schema.CoreSchema:
-        object_id_schema = core_schema.chain_schema(
-            [
-                core_schema.str_schema(),
-                core_schema.no_info_plain_validator_function(cls.validate),
-            ]
-        )
-        return core_schema.json_or_python_schema(
-            json_schema=core_schema.str_schema(),
-            python_schema=core_schema.union_schema(
-                [core_schema.is_instance_schema(ObjectId), object_id_schema]
-            ),
-            serialization=core_schema.plain_serializer_function_ser_schema(
-                str, when_used="json"  # added when_used="json"
-            ),
-        )
-
-    @classmethod
-    def validate(cls, value: Any):
-        if not ObjectId.is_valid(value):
-            raise ValueError("Invalid id")
-
-        return ObjectId(value)
-
-
-PydanticObjectId = Annotated[ObjectId, ObjectIdAnnotation]
+from flaskr.utils import PydanticObjectId
 
 
 class CoreModel(BaseModel):
